Renamed 'codegen_test_util.*' to 'testing_codegen_util.*' to match the test directory naming scheme #690
