+++ test/codegen/group_by_translator_test.cpp
@@ -17,7 +17,7 @@
 #include "expression/conjunction_expression.h"
 #include "planner/aggregate_plan.h"
 
-#include "codegen/codegen_test_util.h"
+#include "codegen/testing_codegen_util.h"
 
 namespace peloton {
 namespace test {
