+++ test/codegen/hash_join_translator_test.cpp
@@ -19,7 +19,7 @@
 #include "planner/hash_plan.h"
 #include "planner/seq_scan_plan.h"
 
-#include "codegen/codegen_test_util.h"
+#include "codegen/testing_codegen_util.h"
 
 namespace peloton {
 namespace test {
