+++ test/codegen/order_by_translator_test.cpp
@@ -15,7 +15,7 @@
 #include "planner/order_by_plan.h"
 #include "planner/seq_scan_plan.h"
 
-#include "codegen/codegen_test_util.h"
+#include "codegen/testing_codegen_util.h"
 
 namespace peloton {
 namespace test {
