+++ test/codegen/value_integrity_test.cpp
@@ -10,7 +10,7 @@
 //
 //===----------------------------------------------------------------------===//
 
-#include "codegen/codegen_test_util.h"
+#include "codegen/testing_codegen_util.h"
 
 #include "codegen/type/tinyint_type.h"
 #include "codegen/type/smallint_type.h"
